Created environnement folder. Refactor file structure.
