+++ src/advanced/tree-sitter/tree_manager.c
@@ -8,9 +8,9 @@
 #include <string.h>
 
 #include "../../../lib/tree-sitter/lib/include/tree_sitter/api.h"
+#include "../../environnement/constants.h"
+#include "../../environnement/global-variables.h"
 #include "../../terminal/highlight.h"
-#include "../../utils/constants.h"
-#include "../../utils/global-variables.h"
 #include "../../utils/tools.h"
 #include "../tree-sitter/tree_query.h"
