+++ src/advanced/tree-sitter/tree_manager.h
@@ -4,8 +4,8 @@
 
 #include "../../../lib/tree-sitter/lib/include/tree_sitter/api.h"
 #include "../../data-management/state_control.h"
+#include "../../environnement/constants.h"
 #include "../../io_management/io_manager.h"
-#include "../../utils/constants.h"
 #include "../theme.h"
 
 #define CHAR_CHUNK_SIZE_TSINPUT 500
