+++ src/advanced/tree-sitter/tree_query.h
@@ -5,7 +5,7 @@
 
 #include "../../../lib/cJSON/cJSON.h"
 #include "../../../lib/tree-sitter/lib/include/tree_sitter/api.h"
-#include "../../utils/constants.h"
+#include "../../environnement/constants.h"
 #include "../../utils/tools.h"
 #include "tree_manager.h"
