+++ src/environnement/global-variables.h
