+++ src/io_management/io_manager.c
@@ -1,5 +1,5 @@
 #include "io_manager.h"
-#include "../utils/constants.h"
+#include "../environnement/constants.h"
 
 #include <assert.h>
 #include <ctype.h>
