+++ src/utils/tools.c
@@ -14,7 +14,7 @@
 #include <sys/time.h>
 #include <unistd.h>
 
-#include "constants.h"
+#include "../environnement/constants.h"
 
 bool areStringEquals(String str1, String str2) { return strcmp(str1.content, str2.content) == 0; }
